feat: move CargoToml struct
